+++ src/components/Others/Snackbar.tsx
@@ -3,7 +3,7 @@ import { useEffect, useLayoutEffect, useRef, useState } from 'react';
 import type { StyleProp, ViewStyle } from 'react-native';
 import { Animated, SafeAreaView, StyleSheet, View } from 'react-native';
 import { Shadow } from 'react-native-shadow-2';
-import { MaterialCommunityIcons } from '@expo/vector-icons';
+import MaterialCommunityIcons from '@expo/vector-icons/MaterialCommunityIcons';
 import { useTheme } from '../../main';
 import { Text } from '../Simple/Text';